Move `SignTransactionDetaillPanel` to `DetailPanel`
