+++ ui/components/SignTransaction/SignTransactionPanelSwitcher.tsx
@@ -1,8 +1,8 @@
 import React, { ReactElement, useState } from "react"
 import { useTranslation } from "react-i18next"
 import SharedPanelSwitcher from "../Shared/SharedPanelSwitcher"
-import SignTransactionDetailPanel from "./SignTransactionDetailPanel"
 import SignTransactionRawDataPanel from "./SignTransactionRawDataPanel"
+import DetailPanel from "../Signing/SignatureDetails/TransactionSignatureDetails/DetailsPanel"
 
 export default function SignTransactionPanelSwitcher(): ReactElement {
   const { t } = useTranslation("translation", { keyPrefix: "signTransaction" })
@@ -15,7 +15,7 @@ export default function SignTransactionPanelSwitcher(): ReactElement {
         panelNumber={panelNumber}
         panelNames={[t("detailPanelName"), t("rawDataPanelName")]}
       />
-      {panelNumber === 0 ? <SignTransactionDetailPanel /> : null}
+      {panelNumber === 0 ? <DetailPanel /> : null}
       {panelNumber === 1 ? <SignTransactionRawDataPanel /> : null}
     </>
   )
